+++ src/test/java/org/thermoweb/intellij/plugin/encrypt/CipherUtilsTest.java
@@ -10,8 +10,8 @@ public class CipherUtilsTest {
 	public void should_encrypt_and_decrypt_string_correctly() {
 		String secretString = "secretString";
 		String password = "password";
-		String encryptedText = CipherUtils.encrypt(secretString, password, Algorithm.PBE.getCode());
-		String decryptedtext = CipherUtils.decrypt(encryptedText, password, Algorithm.PBE.getCode());
+		String encryptedText = CipherUtils.encrypt(secretString, password, Algorithms.PBEWITHHMACSHA1ANDAES_128.getCode());
+		String decryptedtext = CipherUtils.decrypt(encryptedText, password, Algorithms.PBEWITHHMACSHA1ANDAES_128.getCode());
 		assertEquals(secretString, decryptedtext);
 	}