+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 # encrypt-string-plugin Changelog
 
 ## [Unreleased]
+### Added
+- new cipher algorithms
 
 ## [0.1.0] 2022-08-30
 ### Added